+++ ParameterSet/src/Entry.cc
@@ -545,7 +545,8 @@ namespace edm {
 
   bool
   Entry::fromString(std::string::const_iterator const b, std::string::const_iterator const e) {
-    if(static_cast<unsigned long>(e - b) < 4u || b[ 2] != '(' || e[-1] != ')')
+    if(static_cast<unsigned int>(e - b) < 4u || b[ 2] != '(' || e[-1] != ')')
+
       return false;
 
     tracked = b[0];
